[Feat] 투표 기능 구현 및 스크립트 이름 수정
